Account for dolt dsess refactor
